# docs/swagger.yaml
basePath: /
definitions:
  dto.BookingListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.BookingResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.BookingResponse:
    properties:
      created_at:
        type: string
      deposit:
        type: number
      end_date:
        type: string
      id:
        type: string
      monthly_rent:
        type: number
      notes:
        type: string
      owner_id:
        type: string
      rejection_reason:
        type: string
      room_id:
        type: string
      start_date:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateBookingRequest:
    properties:
      end_date:
        description: YYYY-MM-DD
        type: string
      notes:
        maxLength: 500
        type: string
      room_id:
        type: string
      start_date:
        description: YYYY-MM-DD
        type: string
    required:
    - end_date
    - room_id
    - start_date
    type: object
  dto.CreateInvoiceRequest:
    properties:
      booking_id:
        type: string
      due_date:
        description: YYYY-MM-DD
        type: string
      items:
        items:
          $ref: '#/definitions/dto.InvoiceItemRequest'
        minItems: 1
        type: array
      month:
        maximum: 12
        minimum: 1
        type: integer
      notes:
        maxLength: 500
        type: string
      tenant_id:
        type: string
      year:
        minimum: 2020
        type: integer
    required:
    - booking_id
    - due_date
    - items
    - month
    - tenant_id
    - year
    type: object
...
  dto.InvoiceItemRequest:
    properties:
      description:
        type: string
      quantity:
        minimum: 1
        type: integer
      unit_price:
        minimum: 0
        type: number
    required:
    - description
    - quantity
    - unit_price
    type: object
  dto.InvoiceItemResponse:
    properties:
      description:
        type: string
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
  dto.InvoiceListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.InvoiceResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.InvoiceResponse:
    properties:
      booking_id:
        type: string
      created_at:
        type: string
      due_date:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.InvoiceItemResponse'
        type: array
      month:
        type: integer
      notes:
        type: string
      owner_id:
        type: string
      paid_at:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      year:
        type: integer
    type: object
...
  dto.RejectBookingRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    required:
    - reason
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: Production-ready Go REST API starter with hexagonal architecture.
  title: Go Starter API
  version: "1.0"
paths:
...
  /api/v1/bookings:
    get:
      parameters:
      - description: Filter by owner
        in: query
        name: owner_id
        type: string
      - description: Filter by tenant
        in: query
        name: tenant_id
        type: string
      - description: Filter by room
        in: query
        name: room_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingListResponse'
      security:
      - BearerAuth: []
      summary: List bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Create a booking
      tags:
      - bookings
  /api/v1/bookings/{id}:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Get booking by ID
      tags:
      - bookings
  /api/v1/bookings/{id}/activate:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Activate a booking (tenant moved in)
      tags:
      - bookings
  /api/v1/bookings/{id}/approve:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Approve a booking
      tags:
      - bookings
  /api/v1/bookings/{id}/cancel:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - bookings
  /api/v1/bookings/{id}/complete:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Complete a booking (tenant moved out)
      tags:
      - bookings
  /api/v1/bookings/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RejectBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
      security:
      - BearerAuth: []
      summary: Reject a booking
      tags:
      - bookings
  /api/v1/invoices:
    get:
      parameters:
      - description: Filter by booking
        in: query
        name: booking_id
        type: string
      - description: Filter by tenant
        in: query
        name: tenant_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by month (1-12)
        in: query
        name: month
        type: integer
      - description: Filter by year
        in: query
        name: year
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceListResponse'
      security:
      - BearerAuth: []
      summary: List invoices
      tags:
      - invoices
    post:
      consumes:
      - application/json
      parameters:
      - description: Invoice data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Create an invoice
      tags:
      - invoices
  /api/v1/invoices/{id}:
    get:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Get invoice by ID
      tags:
      - invoices
  /api/v1/invoices/{id}/cancel:
    post:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Cancel an invoice
      tags:
      - invoices
  /api/v1/invoices/{id}/download:
    get:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Download invoice as HTML (print-ready)
      tags:
      - invoices
  /api/v1/invoices/{id}/pay:
    post:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Mark invoice as paid
      tags:
      - invoices
  /api/v1/invoices/{id}/send:
    post:
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Send invoice to tenant
      tags:
      - invoices
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"